refactor(workflow): Move `api/core/workflow/workflow_app_generate_task_pipeline.py` to `api/core/app/apps/workflow/generate_task_pipeline.py`
